Move FilePreviewSheet to components/spaces/ for reuse across Dust
