+++ front/components/spaces/FilePreviewSheet.tsx
